aed/requirements.txt
matplotlib==3.0.3
soundfile==0.10.3.post1
librosa==0.6.3
torchlibrosa==0.0.4